Share Redis client helper
